variable id uses unsigned long int.

+++ src/builder/ItemSlotBuilder.hpp
@@ -23,7 +23,7 @@
 #include "../slots/ItemSlot.hpp"
 #include "../common/constants.hpp"
 
-ItemSlot* GetNewItemSlot(int subtype_id, int id = NONE_ID);
+ItemSlot* GetNewItemSlot(int subtype_id, unsigned long int id = NONE_ID);
 ItemSlot* GetNewItemSlotWithoutSaveAbility(int subtype_id);
 
 #endif
